PCHR-3084: Create utils module
