+++ uk.co.compucorp.civicrm.hrcore/gulpfile.js
@@ -8,14 +8,11 @@ var file = require('gulp-file');
 var backstopjs = require('backstopjs');
 var fs = require('fs');
 var path = require('path');
-var PluginError = require('plugin-error');
 var Promise = require('es6-promise').Promise;
-var cv = require('civicrm-cv')({ mode: 'sync' });
 var find = require('find');
 var findUp = require('find-up');
-var xml = require('xml-parse');
 
-var currentExtension;
+var utils = require('./gulp/utils');
 
 // BackstopJS tasks
 (function () {
@@ -169,9 +166,9 @@ var currentExtension;
 
   gulp.task('sass', function (cb) {
     if (hasCurrentExtensionMainSassFile()) {
-      var sequence = addExtensionCustomTasksToSequence([
-        spawnTaskForExtension('sass:sync', syncTask, getCurrentExtension()),
-        spawnTaskForExtension('sass:main', mainTask, getCurrentExtension())
+      var sequence = utils.addExtensionCustomTasksToSequence([
+        utils.spawnTaskForExtension('sass:sync', syncTask, utils.getCurrentExtension()),
+        utils.spawnTaskForExtension('sass:main', mainTask, utils.getCurrentExtension())
       ], 'sass');
 
       gulpSequence.apply(null, sequence)(cb);
@@ -182,8 +179,8 @@ var currentExtension;
   });
 
   gulp.task('sass:watch', function () {
-    var extPath = getExtensionPath();
-    var watchPatterns = addExtensionCustomWatchPatternsToDefaultList([
+    var extPath = utils.getExtensionPath();
+    var watchPatterns = utils.addExtensionCustomWatchPatternsToDefaultList([
       path.join(extPath, 'scss/**/*.scss')
     ], 'sass');
 
@@ -196,7 +193,7 @@ var currentExtension;
    * @return {Boolean}
    */
   function hasCurrentExtensionMainSassFile () {
-    return !!find.fileSync(/\/scss\/([^/]+)?\.scss$/, getExtensionPath())[0];
+    return !!find.fileSync(/\/scss\/([^/]+)?\.scss$/, utils.getExtensionPath())[0];
   }
 
   /**
@@ -206,7 +203,7 @@ var currentExtension;
    * @return {Vinyl}
    */
   function mainTask (cb) {
-    var extPath = getExtensionPath();
+    var extPath = utils.getExtensionPath();
 
     return gulp.src(path.join(extPath, '/scss/*.scss'))
       .pipe(bulk())
@@ -240,14 +237,14 @@ var currentExtension;
   gulp.task('requirejs', function (cb) {
     // The original extension that the task was called with could change during
     // the execution, thus it gets saved so it can be restored later
-    originalExtension = getCurrentExtension();
+    originalExtension = utils.getCurrentExtension();
 
     requireJsTask(cb);
   });
 
   gulp.task('requirejs:watch', function () {
-    var extPath = getExtensionPath();
-    var watchPatterns = addExtensionCustomWatchPatternsToDefaultList([
+    var extPath = utils.getExtensionPath();
+    var watchPatterns = utils.addExtensionCustomWatchPatternsToDefaultList([
       path.join(extPath, '**', 'src/**/*.js')
     ], 'requirejs');
 
@@ -277,18 +274,18 @@ var currentExtension;
     var sequence = buildFiles.filter(function (buildFile) {
       var content = fs.readFileSync(buildFile, 'utf8');
 
-      return (new RegExp(getCurrentExtension(), 'g')).test(content);
+      return (new RegExp(utils.getCurrentExtension(), 'g')).test(content);
     })
       .map(function (buildFileWithDependency) {
-        var extension = getExtensionNameFromFile(buildFileWithDependency);
+        var extension = utils.getExtensionNameFromFile(buildFileWithDependency);
 
-        return spawnTaskForExtension('requirejs', requireJsTask, extension);
+        return utils.spawnTaskForExtension('requirejs', requireJsTask, extension);
       });
 
     sequence.length ? gulpSequence.apply(null, sequence)(function () {
       // Restore the original extension (used in the CLI) as the current extension
       // before marking the task as done
-      setCurrentExtension(originalExtension);
+      utils.setCurrentExtension(originalExtension);
 
       cb();
     }) : cb();
@@ -310,7 +307,7 @@ var currentExtension;
     while ((matches = placeholderRegExp.exec(buildFileContent)) !== null) {
       requiredExtensions.push({
         placeholder: matches[1],
-        path: getExtensionPath(matches[2])
+        path: utils.getExtensionPath(matches[2])
       });
     }
 
@@ -344,7 +341,7 @@ var currentExtension;
    * @param {Function} cb
    */
   function requireJsMainTask (cb) {
-    var buildFilePath = find.fileSync('build.js', getExtensionPath())[0];
+    var buildFilePath = find.fileSync('build.js', utils.getExtensionPath())[0];
     var tempBuildFilePath = path.join(path.dirname(buildFilePath), 'build.tmp.js');
 
     fs.writeFileSync(tempBuildFilePath, processBuildFile(buildFilePath), 'utf8');
@@ -366,13 +363,13 @@ var currentExtension;
     var sequence;
 
     if (!detectInstalled.sync('requirejs')) {
-      throwError('requirejs', 'The `requirejs` package is not installed globally (http://requirejs.org/docs/optimization.html#download)');
+      utils.throwError('requirejs', 'The `requirejs` package is not installed globally (http://requirejs.org/docs/optimization.html#download)');
     }
 
-    sequence = addExtensionCustomTasksToSequence([
-      spawnTaskForExtension('requirejs:main', requireJsMainTask, getCurrentExtension())
+    sequence = utils.addExtensionCustomTasksToSequence([
+      utils.spawnTaskForExtension('requirejs:main', requireJsMainTask, utils.getCurrentExtension())
     ], 'requirejs');
-    sequence.push(spawnTaskForExtension('requirejs:dependencies', extensionDependenciesTask, getCurrentExtension()));
+    sequence.push(utils.spawnTaskForExtension('requirejs:dependencies', extensionDependenciesTask, utils.getCurrentExtension()));
 
     gulpSequence.apply(null, sequence)(cb);
   }
@@ -381,16 +378,16 @@ var currentExtension;
 // Test
 (function () {
   gulp.task('test', function (cb) {
-    var sequence = addExtensionCustomTasksToSequence([
-      spawnTaskForExtension('test:main', mainTask, getCurrentExtension())
+    var sequence = utils.addExtensionCustomTasksToSequence([
+      utils.spawnTaskForExtension('test:main', mainTask, utils.getCurrentExtension())
     ], 'test');
 
     gulpSequence.apply(null, sequence)(cb);
   });
 
   gulp.task('test:watch', function () {
-    var extPath = getExtensionPath();
-    var watchPatterns = addExtensionCustomWatchPatternsToDefaultList([
+    var extPath = utils.getExtensionPath();
+    var watchPatterns = utils.addExtensionCustomWatchPatternsToDefaultList([
       path.join(extPath, '**', 'test/**/*.spec.js'),
       '!' + path.join(extPath, '**', 'test/mocks/**/*.js'),
       '!' + path.join(extPath, '**', 'test/test-main.js')
@@ -421,209 +418,6 @@ var currentExtension;
   });
 }());
 
-/**
- * Given an original sequence of tasks and the name of the "wrapper" task
- * (requirejs, sass, etc), it finds if the current extension
- * has any custom tasks to add before/after or to straight replace the main task
- *
- * @param {Array} sequence
- * @param {String} taskName
- * @return {Array}
- */
-function addExtensionCustomTasksToSequence (sequence, taskName) {
-  var customTasks = getExtensionTasks(taskName);
-
-  if (_.isFunction(customTasks.main)) {
-    var mainIndex = _.findIndex(sequence, function (taskName) {
-      return taskName.match(/:main/);
-    });
-
-    gulp.task(sequence[mainIndex], customTasks.main);
-    sequence.splice(mainIndex, 1, sequence[mainIndex]);
-  }
-
-  if (_.isArray(customTasks.pre)) {
-    customTasks.pre.forEach(function (task, index) {
-      gulp.task(task.name, task.fn);
-      sequence.splice(index, 0, task.name);
-    });
-  }
-
-  if (_.isArray(customTasks.post)) {
-    _.each(customTasks.post, function (task) {
-      gulp.task(task.name, task.fn);
-      sequence.push(task.name);
-    });
-  }
-
-  return sequence;
-}
-
-/**
- * Given a default list of watch patterns and the name of the "main" task
- * (requirejs, sass, etc) if finds if the current extension
- * has any custom task with custom watch patterns to add
- */
-function addExtensionCustomWatchPatternsToDefaultList (defaultList, taskName) {
-  return _(defaultList)
-    .concat(getExtensionTasks(taskName).watchPatterns)
-    .compact()
-    .value();
-}
-
-/**
- * Returns the value of the "key" property of the <extension> tag and of the
- * <file> tag of the given info.xml file
- *
- * @param {String} infoFile
- * @return {Object}
- */
-function getExtensionNameAndAliasFromInfoXML (infoFile) {
-  var parsedXML = xml.parse(fs.readFileSync(infoFile, 'utf8'));
-
-  var extensionTag = _.find(parsedXML, function (node) {
-    return node.tagName && node.tagName === 'extension';
-  });
-
-  return {
-    name: extensionTag.attributes.key,
-    alias: _.find(extensionTag.childNodes, function (node) {
-      return node.tagName && node.tagName === 'file';
-    }).innerXML
-  };
-}
-
-/**
- * Given a file, it finds the info.xml in one of the parent folders and returns
- * the extension name stored in it
- *
- * @param {String} filePath
- * @return {String}
- */
-function getExtensionNameFromFile (filePath) {
-  var infoXMLPath = findUp.sync('info.xml', { cwd: filePath });
-
-  return getExtensionNameAndAliasFromInfoXML(infoXMLPath).name;
-}
-
-/**
- * Given a task name, it looks into the current extension's gulp-task/ folder
- * if there is any file with the name of the task
- */
-function getExtensionTasks (taskName) {
-  var filePath = path.join(getExtensionPath(), '/gulp-tasks/', taskName + '.js');
-
-  if (fs.existsSync(filePath)) {
-    return require(filePath)();
-  } else {
-    return {};
-  }
-}
-
-/**
- * Returns the name of the extension currently used by the tasks
- *
- * If the name is not cached already, it will fetch the name from the CLI argument
- * and then cache it for the next time the function is called
- *
- * @return {String}
- */
-function getCurrentExtension () {
-  if (currentExtension) {
-    return currentExtension;
-  }
-
-  currentExtension = getExtensionNameFromCLI();
-
-  return currentExtension;
-}
-
-/**
- * Returns the extension name specified via CLI argument
- *
- * The name given can be either the real extension name or an alias, which matches
- * the value of the <file> tag in the info.xml file of the extension
- *
- * @return {String}
- * @throws
- *   Will throw an exception in case the argument has not been passed or
- *   no info.xml had been found for the given extension
- */
-function getExtensionNameFromCLI () {
-  var infoFiles, name;
-
-  if (!argv.ext) {
-    throwError('sass', 'Extension name not provided');
-  }
-
-  infoFiles = find.fileSync('info.xml', path.join(__dirname, '..'));
-
-  for (var i = 0; i < infoFiles.length; i++) {
-    var extensionData = getExtensionNameAndAliasFromInfoXML(infoFiles[i]);
-
-    if (extensionData.name === argv.ext || _.endsWith(extensionData.alias, argv.ext)) {
-      name = extensionData.name;
-      break;
-    }
-  }
-
-  if (!name) {
-    throwError('sass', 'Extension "' + argv.ext + '" not found');
-  }
-
-  return name;
-}
-
-/**
- * Uses `cv` to get the path of the given extension
- *
- * @param {String} name If not provided, the name given via the CLI argument is used
- * @return {String}
- */
-function getExtensionPath (name) {
-  var extension = name || getCurrentExtension();
-
-  return cv('path -x ' + extension)[0].value;
-}
-
-/**
- * Sets the given extension as the current one
- *
- * @param {String} extension
- */
-function setCurrentExtension (extension) {
-  currentExtension = extension;
-}
-
-/**
- * Spawns a task for an extension on the fly, using the name and function provided
- * The extension specified is set as the current extension before executing the task fn
- *
- * @param {String} taskName
- * @param {Function} taskFn
- * @param {String} ext
- */
-function spawnTaskForExtension (taskName, taskFn, extension) {
-  taskName += ' (' + extension + ')';
-
-  gulp.task(taskName, function (cb) {
-    setCurrentExtension(extension);
-
-    return taskFn(cb);
-  });
-
-  return taskName;
-}
-
-/**
- * A simple wrapper for displaying errors
- */
-function throwError (plugin, msg) {
-  throw new PluginError('Error', {
-    message: colors.red(msg)
-  });
-}
-
 var test = (function () {
   var find = require('find');
   var karma = require('karma');
@@ -651,7 +445,7 @@ var test = (function () {
   return {
 
     all: function () {
-      var configFile = find.fileSync('karma.conf.js', getExtensionPath())[0];
+      var configFile = find.fileSync('karma.conf.js', utils.getExtensionPath())[0];
 
       runServer(configFile);
     },
